requirements.txt
PyQt6
PyQt6-sip
python-xlib